test: reunite Cache Components adoption eval
