+++ evals/evals/agent-047-adopt-cache-components/PROMPT.md
@@ -1,0 +1,1 @@
+Migrate this storefront to Cache Components incrementally and complete the migration.
