+++ evals/evals/agent-047-adopt-cache-components/app/products/[slug]/page.tsx
