+++ evals/evals/agent-047-adopt-cache-components/lib/catalog.ts
