+++ evals/evals/agent-047-adopt-cache-components/next-env.d.ts
