+++ evals/evals/agent-047-adopt-cache-components/next.config.ts
