+++ evals/evals/agent-047-adopt-cache-components/package.json
