--- evals/evals/agent-047-cache-components-incremental-first-pr/PROMPT.md
@@ -1,1 +1,0 @@
-Migrate this storefront to Cache Components incrementally. Prepare and verify the first migration PR, then stop.
