--- evals/evals/agent-056-cache-components-incremental-completion/PROMPT.md
@@ -1,1 +1,0 @@
-Continue this storefront's Cache Components migration from the first incremental migration PR. Finish the migration on this branch and verify the result.
