--- evals/evals/agent-056-cache-components-incremental-completion/app/privacy/page.tsx
@@ -1,8 +1,0 @@
-export default function PrivacyPage() {
-  return (
-    <main>
-      <h1>Privacy</h1>
-      <p>We only use account data to personalize your experience.</p>
-    </main>
-  )
-}
