--- evals/evals/agent-056-cache-components-incremental-completion/next.config.ts
@@ -1,7 +1,0 @@
-import type { NextConfig } from 'next'
-
-const nextConfig: NextConfig = {
-  cacheComponents: true,
-}
-
-export default nextConfig
